+++ fabric/src/main/java/net/blay09/mods/balm/mixin/PlayerMixin.java
@@ -1,9 +1,11 @@
 package net.blay09.mods.balm.mixin;
 
 import net.blay09.mods.balm.api.Balm;
+import net.blay09.mods.balm.api.entity.BalmPlayer;
 import net.blay09.mods.balm.api.event.DigSpeedEvent;
 import net.blay09.mods.balm.api.event.PlayerAttackEvent;
 import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.Pose;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.block.state.BlockState;
 import org.spongepowered.asm.mixin.Mixin;
@@ -13,7 +15,9 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(Player.class)
-public class PlayerMixin {
+public class PlayerMixin implements BalmPlayer {
+
+    private Pose forcedPose;
 
     @Inject(method = "getDestroySpeed(Lnet/minecraft/world/level/block/state/BlockState;)F", at = @At("RETURN"), cancellable = true)
     private void getDestroySpeed(BlockState state, CallbackInfoReturnable<Float> callbackInfo) {
@@ -38,4 +42,21 @@ public class PlayerMixin {
         }
     }
 
+    @Inject(method = "updatePlayerPose()V", at = @At("HEAD"), cancellable = true)
+    public void updatePlayerPose(CallbackInfo callbackInfo) {
+        if (forcedPose != null) {
+            ((Player) (Object) this).setPose(forcedPose);
+            callbackInfo.cancel();
+        }
+    }
+
+    @Override
+    public Pose getForcedPose() {
+        return forcedPose;
+    }
+
+    @Override
+    public void setForcedPose(Pose pose) {
+        forcedPose = pose;
+    }
 }