+++ shared/java/net/blay09/mods/balm/api/entity/BalmPlayer.java
@@ -1,0 +1,11 @@
+package net.blay09.mods.balm.api.entity;
+
+import net.minecraft.world.entity.Pose;
+
+public interface BalmPlayer {
+
+    Pose getForcedPose();
+
+    void setForcedPose(Pose pose);
+
+}