+++ changelog.md
@@ -1,8 +1,9 @@
 ### Version 2.2.0 for Minecraft 1.18
 
 - Added support for custom farmlands
-- Added support for loot modifiers
 - Added support for custom damage sources
+- Added support for loot modifiers on Fabric
+- Added support for forced poses on Fabric
 - Fixed entity registration on Forge
 
 ### Version 2.1.1 for Minecraft 1.18